fix: 이전 commit으로 checkout을 통해 2주차 반영 내용 삭제하다.
