--- kdt-spring-order/src/main/java/org/prgrms/kdt/entity/Voucher.java
@@ -1,8 +1,0 @@
-package org.prgrms.kdt.entity;
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getVoucherId();
-    String getDiscountValue();
-}
