--- kdt-spring-order/src/main/java/org/prgrms/kdt/repository/VoucherRepository.java
@@ -1,14 +1,0 @@
-package org.prgrms.kdt.repository;
-
-import org.prgrms.kdt.entity.Voucher;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface VoucherRepository {
-    Optional<Voucher> findById(UUID voucherId);
-    void insert (Voucher voucher);
-    List<Voucher> getAllStoredVoucher();
-    void clear();
-}
