+++ src/main/java/org/prgrms/kdt/entity/PercentDiscountVoucher.java
@@ -1,0 +1,31 @@
+package org.prgrms.kdt.entity;
+
+import java.util.UUID;
+
+public class PercentDiscountVoucher implements Voucher {
+    private final UUID voucherId;
+    private final Double percent;
+
+    public PercentDiscountVoucher(UUID voucherId, Double percent) {
+        this.voucherId = voucherId;
+        this.percent = percent;
+    }
+
+    @Override
+    public UUID getVoucherId() {
+        return voucherId;
+    }
+
+    @Override
+    public String getDiscountValue() {
+        return Double.toString(percent);
+    }
+
+    @Override
+    public String toString() {
+        return "PercentDiscountVoucher{" +
+                "voucherId=" + voucherId +
+                ", percent=" + percent +
+                "%}";
+    }
+}
