+++ src/main/java/org/prgrms/kdt/entity/Voucher.java
@@ -1,0 +1,9 @@
+package org.prgrms.kdt.entity;
+
+import java.util.UUID;
+
+public interface Voucher {
+    UUID getVoucherId();
+
+    String getDiscountValue();
+}
