+++ src/main/java/org/prgrms/kdt/voucher/VoucherType.java
