+++ src/main/resources/logback-spring.xml
@@ -1,0 +1,27 @@
+<configuration scan="true">
+    <property name="LOG_PATTERN" value="%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n"/>
+
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>${LOG_PATTERN}</pattern>
+        </encoder>
+    </appender>
+
+    <appender name="ROLLING_FILE_OUT" class="ch.qos.logback.core.rolling.RollingFileAppender">
+        <file>logs/access.log</file>
+        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+            <fileNamePattern>logs/access-%d{yyyy-MM-dd}.log</fileNamePattern>
+        </rollingPolicy>
+        <encoder>
+            <pattern>${LOG_PATTERN}</pattern>
+        </encoder>
+    </appender>
+
+    <!--    <logger name="org.prgrms.kdt" level="warn">-->
+    <!--        <appender-ref ref="ROLLING_FILE_OUT" />-->
+    <!--    </logger>-->
+
+    <root level="warn">
+        <appender-ref ref="ROLLING_FILE_OUT"/>
+    </root>
+</configuration>
